Initial commit
Added project outline, documentation. This will be a long term project,
if it gets started at all. The intention is to design an entire system
capable of running on an FPGA, at a minimum it should be simulated and
could potentially be ported to an FPGA. It will be based on Project
Oberon.