Added vm skeleton, more comments and ideas
Nothing is currently working at the moment, but a plan and skeleton for
the project is forming and this what this commit contains. The license
has also been changed to the 'unlicense', because why not. The project
still needs a name. I am going with 'Howe' for now.

For the VM:

* The instruction set format is nearly decided upon.
* The vast majority of it needs to be implemented.
* This should take a few days of solid work to, and refactor.

Once that VM is a useable state, work will start on the initial cross
compiler that will make the kernel.