New Virtual Machine (design phase)
The design for the Virtual Machine has changed and is much more like the
original computer specified for the Project Oberon FPGA. I do not intend
to replicate the peripherals exactly however. The main differences
between this CPU and original are; Traps/Interrupts, the MMU (TODO), a
modified AND instruction and the fact that the CPU is 64-bit.

There is still much to do.