Test program, MIT License, tracing, MMU/TLB
A lot of the new functionality that has been added has not been tested,
and will still need thinking about, it is likely to be very bug and
non-functional as only a simple test program for the CPU has been added.
The next step in exercising this VM is going to be writing tooling to
exercise things, perhaps a throwaway assembler, but then moving onto the
compiler. It is likely going to be the case that the traps and TLB/MMU
code that has been added will only be fixed much later in the project
once kernel development starts (and those features are actually used).

Changes include:

* Addition of the MIT license for my code
* A small test program that continually prints 'h'.
* Added functions for tracing
* Added more of the TLB functionality, for the moment a MIPs style MMU
(a manual MMU) will be used, this may be changed later.
* A few comments about I/O, floating point numbers, and other stuff has
been added.
* Stack now grows downwards and is initialized to be at the end of
memory by default, which is done in an initialization function.
* More instructions have been added.
* A Real Time Clock has been added.