Added lexer and some parse machinery
The compiler, in 'pc.c', consists of a skeleton of the compiler, and a
working but barely tested lexer. The lexer does seem to produce the
correct output for all the symbols I have tested so far.

The next step is to define the grammar, and more importantly get the
data structures used in the Abstract Syntax Tree correct. This will
define how easy it will be to add more complex things like records,
types and arrays into the compiler.

The work has barely begun.