Added parser, syntax will change
A parser has been added but it will need reworking as the syntax of the
language is decided. It can parse an example program, there are bound to
be errors. The syntax is based off of 'pl/0' and 'C', with some of my
own ideas thrown in, but it is pretty pedestrian at the moment.