Added skeleton for code generation, more grammar
A skeleton for code generation has been added, how it is going to work
has yet to be decided. I think I am going to reuse the AST nodes as much
as possible to avoid new allocations (not on the stack).

I have also added module and import syntax, amongst other things. There
is some type syntax as well.