Removed asm.c, type grammar
The grammar for the language is largely complete, it needs testing out
however. Switch statements are missing, but I can get by without them
for now. It is unclear how I will proceed with many of the tasks that
are left to do, but the next stage is code generation.

Some references to other projects have been added in the 'readme.md'
file.

The throw-away assembler has been thrown away.