Initial commit, skeleton of project
This is the initial commit of the FAT-32 project, it is just an outline
of the project and may not become anything. The idea is to make a
portable FAT-32 implementation and then use that to create a series of
command line utilities that can manipulate FAT-32 images as files on
disk, or as part of a microcontroller. There seems to be a lot of FAT-32
projects on GitHub, but not any that do what I want, they all manage to
get close, but not close enough.