Nearly implemented all of CPU, not bug checked
Most of the meat of the CPU and associated system has been
reimplemented, mostly based on previous work, which could be removed
soon. It is sure to have bugs, especially around the traps and MMU,
which will be ironed out when it is tested. The easiest way to do that
is with a primitive Forth assembler, which will be made in "as.fth".