Added signal register, small test program
The VM is now in a situation where it can be tested more to so what
works and what does not. It appears that the stack machine is a little
to limited in what it can do and might have to be reworked, or not, that
is what the experimentation will be for.